ng-apiselect [enh] split feeder and picker - add feeder component
