+++ packages/ng/src/app/select/api/feeder/index.ts
@@ -1,0 +1,3 @@
+export { ASelectRDDApiFeeder, ISelectApiFeeder } from './api-feeder.model';
+export { RDDApiFeederComponent } from './api-feeder.component';
+export { LuApiFeederModule } from './api-feeder.module';
