+++ packages/ng/src/app/select/api/index.ts
@@ -1,3 +1,4 @@
 export {LuSelectApi} from './select-api.component';
 export { LuSelectApiModule } from './select-api.module';
 export * from './picker';
+export * from './feeder';
