+++ packages/ng/src/app/select/api/picker/index.ts
@@ -1,3 +1,3 @@
-export {ISelectApiFeeder, ASelectRDDApiFeeder} from './select-api-picker.model';
+// export {ISelectApiFeeder, ASelectRDDApiFeeder} from './select-api-picker.model';
 export {LuSelectApiPicker} from './select-api-picker.component';
 export {LuApiPickerModule} from './select-api-picker.module';
